+++ jobs_attributes/alterTable.sql
@@ -1,2 +1,6 @@
-ALTER TABLE /*TABLE_PREFIX*/t_item_job_attr ADD COLUMN i_salary_min_hour INT(6) UNSIGNED NULL DEFAULT NULL  AFTER e_salary_period ,
-ADD COLUMN i_salary_max_hour INT(6) UNSIGNED NULL DEFAULT NULL  AFTER i_salary_min_hour ;
+ALTER TABLE /*TABLE_PREFIX*/t_item_job_attr ADD COLUMN s_salary TEXT NULL DEFAULT '' ;
+ALTER TABLE /*TABLE_PREFIX*/t_item_job_attr DROP i_salary_min ;
+ALTER TABLE /*TABLE_PREFIX*/t_item_job_attr DROP i_salary_max ;
+ALTER TABLE /*TABLE_PREFIX*/t_item_job_attr DROP i_salary_min_hour ;
+ALTER TABLE /*TABLE_PREFIX*/t_item_job_attr DROP i_salary_max_hour ;
+ALTER TABLE /*TABLE_PREFIX*/t_item_job_attr DROP e_salary_period ;