+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog Jobs attributes
 =========================
 
+##version 3.1 - 03/08/2012
+
+* Removed salary slider and chenge it to a free text input
+* Upgrade function
+* To upgrade the plugin, disable and enable it again
+
 ##version 3.0 - 02/05/2012
 
 * Implemented new DAO (Data Access Object)